Move songs query to dedicated file
